+++ client/src/commonMain/kotlin/org/apptank/horus/client/connectivity/ConnectionType.kt
@@ -1,0 +1,6 @@
+package org.apptank.horus.client.connectivity
+
+enum class ConnectionType {
+    WIFI,
+    CELLULAR,
+}