+++ client/src/commonMain/kotlin/org/apptank/horus/client/connectivity/NetworkConnection.kt
@@ -1,0 +1,6 @@
+package org.apptank.horus.client.connectivity
+
+data class NetworkConnection(
+    val type: ConnectionType,
+    val level: ConnectionLevel
+)