+++ client/src/commonMain/kotlin/org/apptank/horus/client/extensions/DatabaseExtension.kt
@@ -53,7 +53,7 @@ internal fun Any?.prepareSQLValueAsString(): String {
 internal fun SqlDriver.createSQLInsert(table: String, values: DataMap): String {
     val columns = values.keys.joinToString(", ")
     val valuesString = values.values.joinToString(", ") { it.prepareSQLValueAsString() }
-    return "INSERT INTO $table ($columns) VALUES ($valuesString)"
+    return "INSERT OR REPLACE INTO $table ($columns) VALUES ($valuesString)"
 }
 
 /**