+++ client/src/commonMain/kotlin/org/apptank/horus/client/sync/network/service/SynchronizationService.kt
@@ -86,6 +86,7 @@ internal class SynchronizationService(
                     downloadedBytes += bytesRead.size
                     sink.write(bytesRead)
                     contentLength?.let {
+                        info("[SynchronizationService] Download progress: $downloadedBytes / $contentLength")
                         onProgress(((downloadedBytes / contentLength.toDouble()) * 100).toInt())
                     }
                 }
@@ -144,7 +145,7 @@ internal class SynchronizationService(
      */
     override suspend fun postQueueActions(actions: List<SyncDTO.Request.SyncActionRequest>): DataResult<Unit> {
 
-        val chunks = actions.sortedBy { it.actionedAt }.chunked(1000)
+        val chunks = actions.sortedBy { it.actionedAt }.chunked(500)
         val results = mutableListOf<DataResult<Unit>>()
 
         chunks.forEach { chunk ->